Initial implementation of slab pmem
The datapool_pmem implementation, which can be enabled through
-DUSE_PMEM=ON, allows for durable (persistent) implementations of the
storage layer as it attempts to recover the previous content
of the managed memory.

This patch uses the datapool interface to allow for durable
operations of the slab storage engine.

To enable durable storage, provide a path to a file in
"slab_datapool" config option. If no file is provided, the
datapool falls back to shared memory.

* Cmake workaround for fix check detection introduced in
 f7d34f47c86a3761ac537d16fea43855590c8c96