Slab pmem implementation
This could be enabled through-DUSE_PMEM=ON, allows for durable
(persistent) implementations of the storage layer as it attempts to
recover the previous content of the managed memory.

This patch uses the datapool interface to allow for durable
operations of the slab storage engine.

To enable durable storage, provide a path to a file in "slab_datapool"
config option. If no file is provided, the datapool falls back to shared
memory.